implement posting of thoughts
