+++ src/Thought/Thought.jsx
